Merges tests in to one file.
